+++ frontend/src/main.tsx
@@ -3,11 +3,14 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
 import { ThemeProvider } from './theme.tsx';
+import { LocaleProvider } from './i18n/locale.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <ThemeProvider>
-      <App />
-    </ThemeProvider>
+    <LocaleProvider>
+      <ThemeProvider>
+        <App />
+      </ThemeProvider>
+    </LocaleProvider>
   </StrictMode>,
 );